Fixed a crash caused by camera device switch resulting in undef format wchi is -1, but the parameter was type uint16_t which was not handled correctly.

+++ src/video/camerainfo.h
@@ -22,7 +22,7 @@ public:
                      QSize& resolution, double& framerate, QString &formatText);
 
   // used by dshow camera to determine the index of capability
-  int formatToCapability(uint16_t device, uint16_t formatIndex, uint16_t resolutionIndex);
+  int formatToCapability(uint16_t device, int formatIndex, uint16_t resolutionIndex);
 
 private:
